+++ autoTradingSystem/src/__init__.py
@@ -2,4 +2,4 @@
 Source Package
 """
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'